requirements.txt
easyocr
customtkinter
pillow==9.5.0
pyttsx3